feat(ffi-toolkit): add response status enum for FFI
The `FCPResponseStatus` enum used to be part of the sector builder. It is
now moved to ffi-toolkit so that other Filecoin FFIs can use the exact
same response status codes.

The name of the enum is kept, so that FFIs using it, will still produce
the same header files when cbindgen is used.

Part of https://github.com/filecoin-project/rust-fil-proofs-ffi/issues/8.