feat(ffi-toolkit): add closure to catch panics
When panics happen on the Rust side of the FFI, normally it just aborts
and the caller has no chance to catch the error.

When a a FFI function is wrapped in a `catch_panic_response()` closure,
then whenever a panic (where unwinding is possible) occurs it will respond
to the caller with the `FCPResponseStatus::FCPUnclassifiedError` error code
and the error message prefixed with "Rust panic: ". Now the caller can
react accordingly.

This is part of https://github.com/filecoin-project/rust-fil-proofs/issues/553