chore: use --locked instead of --frozen
`--frozen` and `--locked` are very similar. They share the idea to
use the same existing lockfile. One difference we make use of here
is, that `--locked` will also work if the Rust version used to create
the lock file is different from the one that is used to build/test
the project.

This is needed for this CI config, as the Rust version bundled with
the image is way older than the version specified in `rust-toolchain`.

Also don't persist things that are not re-used.